Updated everything

+++ P_FOLDER/p.c
@@ -67,11 +67,11 @@ void findTime(int n,int bt[],int p[]){
         printf("%13d",tat[i] ); 
         printf("\n");
 	}
-	int s=(float)total_wt / (float)n; 
-    int t=(float)total_tat / (float)n; 
-    printf("Average waiting time = %d",s); 
+	float s=total_wt / (float)n; 
+    float t=total_tat / (float)n; 
+    printf("Average waiting time = %.2f",s); 
     printf("\n"); 
-    printf("Average turn around time = %d ",t);
+    printf("Average turn around time = %.2f ",t);
 }
 int main(int argc, char const *argv[]){
 	int i=0;
